Funciona - Mapa reducido

+++ src/actions/IrOeste.java
@@ -25,7 +25,7 @@ public class IrOeste extends SearchAction{
 
 	@Override
 	public Double getCost() {
-		return new Double(1);
+		return new Double(2);
 	}
 
 	@Override
