+++ src/actions/IrSur.java
@@ -25,7 +25,7 @@ public class IrSur extends SearchAction{
 
 	@Override
 	public Double getCost() {
-		return new Double(1);
+		return new Double(2);
 	}
 
 	@Override
